### 2025-10-13 #95 @nyida
[Re] Watts–Strogatz Small-World Networks
**Original article:**
Watts, D. J., & Strogatz, S. H. (1998). Collective dynamics of ‘small-world’ networks. Nature, 393(6684), 440–442.

**PDF URL:**
https://zenodo.org/records/17336306/files/ReScienceC.pdf

**Scientific domain:** Network science; complex systems; computational modeling
