### 2025-10-20 #95 @rougier
Thanks for your submission. Can you import your notebook in the repository (and ideally, it would be better to have simple Python scripts for long time archiving)?
