+++ [re]_watts–strogatz_small_world_networks.py
@@ -1,0 +1,327 @@
+# -*- coding: utf-8 -*-
+"""[Re] Watts–Strogatz Small-World Networks
+
+Automatically generated by Colab.
+
+Original file is located at
+    https://colab.research.google.com/drive/1IdA9jBV-BgWIZxIbn-LF1a41U49X0MMN
+"""
+
+# ============================================================
+# [Re] Watts–Strogatz Small-World Networks
+# - CONSISTENT METRICS on giant component (Gc)
+# - GC FILTER (φ>=0.95) for apples-to-apples normalized C/L curves
+# - FULL-GRAPH ROBUSTNESS: global efficiency E(G) across all p
+# - THEORY + EMPIRICAL random baselines for C and L
+# - REWIRING VALIDATION: lattice-edge fraction vs p
+# - DEGREE + ASSORTATIVITY diagnostics vs p
+# - 50 seeds + CI stability analysis (L at p=1e-2)
+# - Outputs: CSVs, PNGs, metadata.json, and a printed validation report
+# ============================================================
+
+!pip install -q networkx==3.3
+
+import os, sys, json, time, math, csv, random
+from pathlib import Path
+import numpy as np
+import matplotlib.pyplot as plt
+import networkx as nx
+
+# -------- Determinism --------
+os.environ["PYTHONHASHSEED"] = "0"
+random.seed(0)
+np.random.seed(0)
+
+# -------- Paths --------
+ROOT = Path("results"); FIGS = ROOT/"figs"; DATA = ROOT/"data"
+for d in (FIGS, DATA): d.mkdir(parents=True, exist_ok=True)
+
+# -------- Experiment params --------
+N = 1000              # nodes
+k = 10                # mean degree (must be even)
+assert k % 2 == 0
+# p-grid: 0 plus logspaced 1e-4..1
+p_values = np.unique(np.concatenate([[0.0], np.round(np.logspace(-4, 0, 13), 4)]))
+n_seeds = 50          # statistical power for 95% CIs
+MIN_GC_FRAC = 0.95    # apples-to-apples threshold for normalized curves
+
+# -------- Helpers --------
+def mean_ci95(x):
+    x = np.asarray(x, float)
+    m = x.mean()
+    if len(x) < 2:
+        return float(m), 0.0
+    s = x.std(ddof=1) / max(len(x)**0.5, 1e-12)
+    return float(m), float(1.96 * s)
+
+def ws_graph(N, k, p, seed):
+    return nx.watts_strogatz_graph(N, k, float(p), seed=seed)
+
+def to_giant_component_view(G):
+    if nx.is_connected(G):
+        return G, 1.0
+    nodes = max(nx.connected_components(G), key=len)
+    Gc = G.subgraph(nodes)  # view, no copy
+    return Gc, Gc.number_of_nodes() / G.number_of_nodes()
+
+def graph_metrics_consistent(G):
+    Gc, frac = to_giant_component_view(G)
+    C = nx.average_clustering(Gc)
+    L = nx.average_shortest_path_length(Gc)
+    return C, L, frac, Gc
+
+def degree_stats(G):
+    deg = np.array([d for _, d in G.degree()], dtype=float)
+    return float(deg.mean()), float(deg.var())
+
+def assortativity(G):
+    try:
+        return float(nx.degree_assortativity_coefficient(G))
+    except Exception:
+        return float('nan')
+
+def ring_edges_set(N, k):
[…]
