### 2025-11-28 #95 @nyida
Hi,

Thank you for the feedback and for reviewing my submission.
I wanted to let you know that I’ve now added:

the full Jupyter notebook (.ipynb)

a simplified Python script version (.py)

Both are included in the repository as requested.
Please let me know if you'd like any additional changes or adjustments.

Thanks again for your time and for guiding the replication process

Best,
Nyida
