### 2025-12-04 #95 @rougier
Sorry for all these delays. Things should be smoother from now on.
I did not find the link to your PDF article. Can you edit your issue and add the link?
