### 2025-12-13 #95 @nyida
Hi, I’ve edited the issue to include the direct link to the PDF article on Zenodo.
