### 2026-01-08 #95 @rougier
Thanks. But you'll have to use the template. Please have a look at https://rescience.github.io/write/
