### 2026-03-04 98a33dd
Create draft.tex
