+++ draft.tex
@@ -1,0 +1,12 @@
+\documentclass{rescience}
+
+\input{metadata.tex}
+\input{header.tex}
+
+\begin{document}
+
+\input{content.tex}
+
+\bibliography{bibliography}
+
+\end{document}
