### 2026-03-04 f45f837
Update bibliography
