+++ bibliography.bib
@@ -1,88 +1,79 @@
-@article{Rougier:2018,
-  doi = {10.1038/d41586-018-04628-w},
-  year = {2018},
-  month = apr,
-  publisher = {Springer Nature},
-  volume = {556},
-  number = {7701},
-  pages = {309--309},
-  author = {Rougier, Nicolas P. and Hinsen, Konrad},
-  title = {Code reviewing puts extra demands on referees },
+@article{WattsStrogatz1998,
+  author  = {Watts, Duncan J. and Strogatz, Steven H.},
+  title   = {Collective dynamics of 'small-world' networks},
   journal = {Nature},
+  volume  = {393},
+  number  = {6684},
+  pages   = {440--442},
+  year    = {1998},
+  doi     = {10.1038/30918}
 }
 
-@article{Science:2018,
-  doi = {10.1126/science.359.6377.725},
-  author = {Hutson, Matthew},
-  url = {http://science.sciencemag.org/content/359/6377/725},
-  year = {2018},
-  month = feb,
-  volume = {359},
-  number = {6377},
-  title = {Artificial intelligence faces a replication crisis},
-  journal = {Science},
+@article{Rougier2018,
+  doi       = {10.1038/d41586-018-04628-w},
+  year      = {2018},
+  month     = {apr},
+  publisher = {Springer Nature},
+  volume    = {556},
+  number    = {7701},
+  pages     = {309--309},
+  author    = {Rougier, Nicolas P. and Hinsen, Konrad},
+  title     = {Code reviewing puts extra demands on referees},
+  journal   = {Nature}
 }
 
-@Misc{Roboto:2011,
-  author =       {Christian Robertson},
-  title =        {The Roboto family of fonts (Google)},
-  url =          {https://github.com/google/roboto},
-  year =         2011,
-  note =         {Apache License, verison 2.0},
+@article{Hutson2018,
+  doi       = {10.1126/science.359.6377.725},
+  author    = {Hutson, Matthew},
+  url       = {http://science.sciencemag.org/content/359/6377/725},
+  year      = {2018},
+  month     = {feb},
+  volume    = {359},
+  number    = {6377},
+  title     = {Artificial intelligence faces a replication crisis},
+  journal   = {Science}
 }
 
-@Misc{SourceSerifPro:2014,
-  author =       {Frank Grießhammer},
-  title =        {Source Serif Pro (Adobe Systems)},
-  url =          {https://github.com/adobe-fonts/source-serif-pro},
-  year =         2014,
-  note =         {SIL Open Font License, version 1.1},
+@misc{Roboto2011,
+  author = {Christian Robertson},
+  title  = {The Roboto family of fonts (Google)},
+  url    = {https://github.com/google/roboto},
+  year   = {2011},
+  note   = {Apache License, version 2.0}
 }
 
-@Misc{SourceCodePro:2012,
-  author =       {Paul D. Hunt},
-  title =        {Source Code Pro (Adobe Systems)},
-  url =          {https://github.com/adobe-fonts/source-code-pro},
-  year =         2012,
-  note =         {SIL Open Font License, version 1.1},
+@misc{SourceSerifPro2014,
+  author = {Frank Grießhammer},
+  title  = {Source Serif Pro (Adobe Systems)},
+  url    = {https://github.com/adobe-fonts/source-serif-pro},
+  year   = {2014},
+  note   = {SIL Open Font License, version 1.1}
 }
 
-@article{Topalidou:2015,
-  author =       {Topalidou, Meropi and Rougier, Nicolas P.},
-  title =        {{[Re] Interaction between cognitive and motor cortico-basal
-                  ganglia loops during decision making: a computational study}},
[…]
