### 2026-03-04 #95 @nyida
Hi, I have updated the repository to use the official ReScience template and added the required template files.
