### 2026-03-09 #95 @rougier
Thanks. I've read your replication but from what I understood, it is mainly based on networkX Watts Strogatz graph implementation, right? It is thuis not clear what is the added value of your code? Could you develop a bit?
